remove unuse code after plan2 refactor
