+++ pkg/sql/plan2/types.go
@@ -77,28 +77,6 @@ type BaseOptimizer struct {
 	ctx   CompilerContext
 }
 
-//use for build select
-type BinderContext struct {
-	// when build_projection we may set columnAlias and then use in build_orderby
-	columnAlias map[string]*Expr
-	// when build_cte will set cteTables and use in build_from
-	cteTables map[string]*TableDef
-
-	// use for build subquery
-	subqueryIsCorrelated bool
-	// unused, commented out for now.
-	// subqueryIsScalar     bool
-
-	subqueryParentIds []int32
-
-	// use to storage the using columns.
-	// select R.*, S.* from R, S using(a) where S.a > 10
-	// then we store {'a':'S'},
-	// when we use buildUnresolvedName(), and the colName = 'a' and tableName = 'S', we reset tableName=''
-	// because the ProjectNode(after JoinNode) had coalesced the using cols
-	usingCols map[string]string
-}
-
 ///////////////////////////////
 // Data structures for refactor
 ///////////////////////////////
